Nick's homepage now shows up at the index. Also renamed some things for readability
